// src/app/model/transaction.ts
export class Transaction {
    purpose: string | undefined;
    mode: string | undefined;
    givenAmount: number | undefined;
    receivedAmount: number | undefined;
    netClosingAmount: number | undefined;
    type: string | undefined;
    givenBy: string | undefined;
    takenBy: string | undefined;
    createdDate: string | undefined;
    updatedDate: string | undefined;
    category: string | undefined;
}